feat: add trigger and icon for changing theme, refactor bloc folder structure
