+++ lib/module/calculator/screen/calculator_screen.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
-import '../bloc/calculator_bloc.dart';
+import 'package:glassmorphism_ui/glassmorphism_ui.dart';
+import 'package:my_calculator/theme.dart';
+import '../blocs/calculator_bloc/calculator_bloc.dart';
 import '../section/display.dart';
 import '../section/keyboard.dart';
 
@@ -15,27 +16,30 @@ class CalculatorScreen extends StatefulWidget {
 class _CalculatorScreenState extends State<CalculatorScreen> {
   _onPressed(String command) {
     setState(() {
-      context.read<CalculatorBloc>().add(CalculatorEvent(command: command));
+      // context.read<CalculatorBloc>().add(CalculatorEvent(command: command));
     });
   }
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: Container(
+    return Scaffold(
+      body: Container(
         decoration: new BoxDecoration(
             image: new DecorationImage(
-          image: new AssetImage("assets/images/bg_white.png"),
+          image: new AssetImage("assets/images/bg.png"),
           fit: BoxFit.fill,
         )),
-        child: Column(
-          children: <Widget>[
-            BlocBuilder<CalculatorBloc, CalculatorState>(builder: (__, state) {
-              return Display(state.value);
-            }),
-            Keyboard(_onPressed),
-          ],
+        child: GlassContainer(
+          color: kBlackColor,
+          child: Column(
+            children: <Widget>[
+              BlocBuilder<CalculatorBloc, CalculatorState>(
+                  builder: (__, state) {
+                return Display(state.value);
+              }),
+              Keyboard(_onPressed),
+            ],
+          ),
         ),
       ),
     );
