+++ lib/module/calculator/section/display.dart
@@ -1,38 +1,56 @@
 import 'package:flutter/material.dart';
 import 'package:auto_size_text/auto_size_text.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:glassmorphism_ui/glassmorphism_ui.dart';
+import 'package:my_calculator/theme.dart';
 
-class Display extends StatelessWidget {
+class Display extends StatefulWidget {
   final String text;
   Display(this.text);
 
   @override
+  State<Display> createState() => _DisplayState();
+}
+
+bool _iconBool = false;
+IconData _iconLight = Icons.wb_sunny;
+IconData _iconDark = Icons.nights_stay;
+
+class _DisplayState extends State<Display> {
+  @override
   Widget build(BuildContext context) {
     return Expanded(
       flex: 1,
-      child: Container(
-        color: Colors.transparent,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.end,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: <Widget>[
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: AutoSizeText(
-                text,
-                minFontSize: 20,
-                maxFontSize: 80,
-                maxLines: 1,
-                textAlign: TextAlign.end,
-                style: TextStyle(
-                  fontWeight: FontWeight.w100,
-                  decoration: TextDecoration.none,
-                  fontSize: 80,
-                  color: Colors.white,
-                ),
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.end,
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: <Widget>[
+          IconButton(
+            onPressed: () {
+              //  context.read<CalculatorBloc>().add(CalculatorEvent(command: command));
+              setState(() {
+                _iconBool = !_iconBool;
+              });
+            },
+            icon: Icon(_iconBool ? _iconDark : _iconLight),
+          ),
+          Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: AutoSizeText(
+              widget.text,
+              minFontSize: 20,
+              maxFontSize: 80,
+              maxLines: 1,
+              textAlign: TextAlign.end,
+              style: TextStyle(
+                fontWeight: FontWeight.w100,
+                decoration: TextDecoration.none,
+                fontSize: 80,
+                color: Colors.white,
               ),
             ),
-          ],
-        ),
+          ),
+        ],
       ),
     );
   }
